+++ internal/vehicles/domain/entities/sell_vehicle.go
@@ -1,0 +1,7 @@
+package entities
+
+type SellVehicle struct {
+	VehicleID string  `json:"vehicle_id"`
+	Cpf       string  `json:"cpf"`
+	Amount    float64 `json:"amount"`
+}